Add item 3.2.3 - capitalization of acronyms, common acronyms and logograms
"Refer typographic disputes to the higher courts of speech and thinking"

This one is pretty thin as there's not much CSS can do without knowing the
semantics of the text and even the particular state of a given acronym in the
popular imagination. But just skipping a section felt wrong too (?) so here it
is.